+++ AlmacenApi.Application/Features/Combo/Dto/ProductsComboDto.cs
@@ -2,5 +2,5 @@ namespace AlmacenApi.Aplication.Features.Combo.Dto;
 public class ProductComboDto
 {
     public required string Id{ get ; set ;}
-    public required int Quantity { get ; set ;}
+    public required double Quantity { get ; set ;}
 }
